refactor : 1. 계산 기능을 외부로 클래스 메서드로 분리 2. 일부 클래스명 변경하여 역할 명시
