+++ src/test/java/calculator/CalculatorTest.java
@@ -3,18 +3,10 @@ package calculator;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 class CalculatorTest {
 
-//    "2 + 3 * 4 / 2"
-//    @ParameterizedTest
-//    @CsvSource(value = {"2 + 3:'5'", "2 + 3 * 4:20", "2 + 3 * 4 / 2:10"}, delimiter = ':')
-    
     @Test
     void calculateAddTest() {
         Calculator calculator = new Calculator(new StringInput("2 + 3"));
