+++ src/test/java/calculator/prepare/AccumulatorTest.java
@@ -1,0 +1,26 @@
+package calculator.prepare;
+
+import static calculator.prepare.Accumulator.accumulateNumber;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class AccumulatorTest {
+    
+    @Test
+    void addTest() {
+        assertThat(accumulateNumber("+", 5, 5)).isEqualTo(10);
+    }
+    
+    @Test
+    void subTest() {
+        assertThat(accumulateNumber("-", 5, 0)).isEqualTo(5);
+    }
+    
+    @Test
+    void OperatorSetTest() {
+        assertThat(accumulateNumber("*", 5, 3)).isEqualTo(15);
+    }
+    
+}
